Add mzmine 3 test data, rename methods
